requirements_lora_vla.txt
# LoRA-Enhanced VLA System Requirements
# Optimized for 6GB GPU training and comprehensive evaluation

# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and Model Libraries
transformers>=4.30.0
tokenizers>=0.13.0
accelerate>=0.20.0

# Parameter-Efficient Fine-Tuning
peft>=0.4.0
bitsandbytes>=0.39.0

# Data Processing and Utilities
numpy>=1.21.0
pandas>=1.3.0
pillow>=9.0.0
datasets>=2.0.0
scikit-learn>=1.0.0

# Visualization and Plotting
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress and CLI
tqdm>=4.64.0
rich>=12.0.0
typer>=0.6.0

# Configuration and Serialization
pyyaml>=6.0
toml>=0.10.0
omegaconf>=2.2.0

# File and Data Handling
h5py>=3.7.0
pickle5>=0.0.11; python_version < '3.8'

# Image and Video Processing
opencv-python>=4.6.0
imageio>=2.19.0
imageio-ffmpeg>=0.4.7

# Scientific Computing
scipy>=1.8.0
einops>=0.6.0

# Logging and Monitoring
wandb>=0.13.0
tensorboard>=2.10.0

# Development and Testing
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0

# Jupyter and Interactive Development
jupyter>=1.0.0
ipywidgets>=7.7.0
notebook>=6.4.0

# Model Utilities
safetensors>=0.3.0
huggingface-hub>=0.15.0

# Additional MetaWorld Dependencies (if using real environments)
# mujoco>=2.3.0
# gym>=0.21.0
# metaworld>=2.0.0

# Memory Profiling and Optimization
memory-profiler>=0.60.0
psutil>=5.9.0

# Async and Parallel Processing
aiofiles>=0.8.0
joblib>=1.1.0

# Network and HTTP
requests>=2.28.0
urllib3>=1.26.0

# Time and Date Utilities
python-dateutil>=2.8.0

# System and Environment
packaging>=21.0
importlib-metadata>=4.0.0; python_version < '3.8'

# GPU Memory Management
nvidia-ml-py3>=7.352.0

# Additional utilities for robotic simulation
# pybullet>=3.2.0  # Alternative physics engine
# dm-control>=1.0.0  # DeepMind control suite

# Optional: For advanced model serving
# fastapi>=0.68.0
# uvicorn>=0.15.0

# Optional: For distributed training
# deepspeed>=0.7.0

# Version constraints for compatibility
# Pin critical versions to avoid conflicts
# torch-audio==2.0.*
# torch-vision==0.15.* 